+++ src/Platform/Dashboard.php
@@ -11,7 +11,7 @@ class Dashboard
     /**
      * ORCHID Version.
      */
-    public const VERSION = '3.10.5';
+    public const VERSION = '3.10.6';
 
     /**
      * The Dashboard configuration options.